Split up evidence command module into its own tab
